requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0
slowapi>=0.1.9
groq>=0.9.0
supabase>=2.0.0
httpx>=0.27.0